requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.10
PyJWT==2.10.1
python-dotenv==1.0.0
pytz==2024.2
requests==2.31.0
sqlparse==0.5.2
urllib3==2.2.3
